backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator>=2.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pymysql==1.1.0
cryptography==41.0.7
python-dotenv==1.0.0
requests>=2.31.0


chromadb==0.4.22
dashscope>=1.17.0
langchain>=0.1.0
langchain-community>=0.0.20


PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0

httpx>=0.22.0,<0.25.0

numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.8.0

